+++ src/main/java/me/hydos/lint/block/DirtLikeBlock.java
@@ -17,24 +17,33 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package me.hydos.lint.refactord.block.organic;
+package me.hydos.lint.block;
 
+import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.SaplingBlock;
-import net.minecraft.block.sapling.SaplingGenerator;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.BlockView;
+import net.minecraft.block.Blocks;
 
-public class LintSaplingBlock extends SaplingBlock {
-	private final BlockState requires;
-
-	public LintSaplingBlock(SaplingGenerator generator, Settings settings, BlockState requires) {
-		super(generator, settings);
-		this.requires = requires;
+public class DirtLikeBlock extends Block {
+	public DirtLikeBlock(Settings settings) {
+		super(settings);
 	}
 
 	@Override
-	protected boolean canPlantOnTop(BlockState floor, BlockView view, BlockPos pos) {
-		return floor == this.requires;
+	public boolean is(Block block) {
+		return this == block || Blocks.DIRT == block;
+	}
+
+	public static boolean isLintGrass(BlockState state) {
+		Block block = state.getBlock();
+		return isLintGrass(block);
+	}
+
+	public static boolean isLintGrass(Block block) {
+		return block == LintBlocks.CORRUPT_GRASS || block == LintBlocks.LIVELY_GRASS || block == LintBlocks.FROSTED_GRASS;
+	}
+
+	public static boolean isUntaintedGrass(BlockState state) {
+		Block block = state.getBlock();
+		return block == LintBlocks.LIVELY_GRASS || block == LintBlocks.FROSTED_GRASS;
 	}
 }
