+++ src/main/java/me/hydos/lint/block/organic/FallenLeavesBlock.java
@@ -17,35 +17,24 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package me.hydos.lint.refactord.block.organic;
+package me.hydos.lint.block.organic;
 
-import me.hydos.lint.refactord.block.LintBlocks2;
-import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
+import net.minecraft.block.LeavesBlock;
 import net.minecraft.block.ShapeContext;
-import net.minecraft.block.TallFlowerBlock;
-import net.minecraft.block.enums.DoubleBlockHalf;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.util.shape.VoxelShapes;
 import net.minecraft.world.BlockView;
 
-public class LintTallFlowerBlock extends TallFlowerBlock {
-	public static final VoxelShape BOTTOM_MODEL = VoxelShapes.cuboid(0.125, 0, 0.125, 0.875, 1, 0.875);
-	public static final VoxelShape TOP_MODEL = VoxelShapes.cuboid(0.125, 0, 0.125, 0.875, 0.75, 0.875);
+public class FallenLeavesBlock extends LeavesBlock {
 
-	public LintTallFlowerBlock(Settings settings) {
+	public FallenLeavesBlock(Settings settings) {
 		super(settings);
 	}
 
 	@Override
-	protected boolean canPlantOnTop(BlockState floor, BlockView world, BlockPos pos) {
-		Block block = floor.getBlock();
-		return block == LintBlocks2.LIVELY_GRASS || block == LintBlocks2.CORRUPT_GRASS;
-	}
-
-	@Override
-	public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
-		return state.get(HALF) == DoubleBlockHalf.LOWER ? BOTTOM_MODEL : TOP_MODEL;
+	public VoxelShape getOutlineShape(BlockState state, BlockView view, BlockPos pos, ShapeContext context) {
+		return VoxelShapes.cuboid(0f, 0f, 0f, 1f, 0.0625f, 1f);
 	}
 }
