+++ src/main/java/me/hydos/lint/block/organic/LintFlowerBlock.java
@@ -17,11 +17,11 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package me.hydos.lint.refactord.block.organic;
+package me.hydos.lint.block.organic;
 
 import org.jetbrains.annotations.Nullable;
 
-import me.hydos.lint.refactord.block.LintBlocks2;
+import me.hydos.lint.block.LintBlocks;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.FlowerBlock;
 import net.minecraft.block.ShapeContext;
@@ -46,7 +46,7 @@ public class LintFlowerBlock extends FlowerBlock {
 
 	@Override
 	protected boolean canPlantOnTop(BlockState floor, BlockView view, BlockPos pos) {
-		return floor == LintBlocks2.LIVELY_GRASS.getDefaultState();
+		return floor == LintBlocks.LIVELY_GRASS.getDefaultState();
 	}
 
 	@Override
