+++ src/main/java/me/hydos/lint/block/organic/LintSpreadableBlock.java
@@ -17,11 +17,11 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package me.hydos.lint.refactord.block.organic;
+package me.hydos.lint.block.organic;
 
 import java.util.Random;
 
-import me.hydos.lint.refactord.block.LintBlocks2;
+import me.hydos.lint.block.LintBlocks;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.block.SnowBlock;
@@ -59,14 +59,14 @@ public class LintSpreadableBlock extends SnowyBlock {
 	@Override
 	public void randomTick(BlockState state, ServerWorld world, BlockPos pos, Random random) {
 		if (!canSurvive(state, world, pos)) {
-			world.setBlockState(pos, LintBlocks2.RICH_DIRT.getDefaultState());
+			world.setBlockState(pos, LintBlocks.RICH_DIRT.getDefaultState());
 		} else {
 			if (world.getLightLevel(pos.up()) >= 9) {
 				BlockState blockState = this.getDefaultState();
 
 				for(int i = 0; i < 4; ++i) {
 					BlockPos setPos = pos.add(random.nextInt(3) - 1, random.nextInt(5) - 3, random.nextInt(3) - 1);
-					if (world.getBlockState(setPos).isOf(LintBlocks2.RICH_DIRT) && canSpread(blockState, world, setPos)) {
+					if (world.getBlockState(setPos).isOf(LintBlocks.RICH_DIRT) && canSpread(blockState, world, setPos)) {
 						world.setBlockState(setPos, blockState.with(SNOWY, world.getBlockState(setPos.up()).isOf(Blocks.SNOW)));
 					}
 				}
