+++ src/main/java/me/hydos/lint/block/util/BlockBuilder.java
@@ -17,7 +17,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package me.hydos.lint.block;
+package me.hydos.lint.block.util;
 
 import me.hydos.lint.Lint;
 import me.hydos.lint.item.group.ItemGroups;
