+++ src/main/java/me/hydos/lint/fluid/LintFluids.java
@@ -20,7 +20,7 @@
 package me.hydos.lint.fluid;
 
 import me.hydos.lint.Lint;
-import me.hydos.lint.refactord.block.LintBlocks2;
+import me.hydos.lint.block.LintBlocks;
 import net.minecraft.block.Blocks;
 import net.minecraft.fluid.FlowableFluid;
 import net.minecraft.fluid.FluidState;
@@ -39,7 +39,7 @@ public class LintFluids {
     public static void initialize() {
         MOLTEN_FLUID_MAP.put(Lint.id("iron"), new FluidEntry(Items.IRON_NUGGET, Items.IRON_INGOT, Blocks.IRON_BLOCK.asItem(), "iron", 0x44FFFFFF));
         MOLTEN_FLUID_MAP.put(Lint.id("gold"), new FluidEntry(Items.GOLD_NUGGET, Items.GOLD_INGOT, Blocks.GOLD_BLOCK.asItem(), "gold", 0x44ffdd30));
-        MOLTEN_FLUID_MAP.put(Lint.id("lava"), new FluidEntry(Items.AIR, Items.LAVA_BUCKET, LintBlocks2.FUSED_STONE.asItem(), "lava", 0xFFFFFFFF));
+        MOLTEN_FLUID_MAP.put(Lint.id("lava"), new FluidEntry(Items.AIR, Items.LAVA_BUCKET, LintBlocks.FUSED_STONE.asItem(), "lava", 0xFFFFFFFF));
     }
 
     public static FluidEntry get(String path) {
