+++ src/main/java/me/hydos/lint/item/WaypointTeleportItem.java
@@ -19,8 +19,8 @@
 
 package me.hydos.lint.item;
 
-import me.hydos.lint.refactord.block.LintBlocks2;
-import me.hydos.lint.refactord.block.ReturnHomeBlock;
+import me.hydos.lint.block.LintBlocks;
+import me.hydos.lint.block.ReturnHomeBlock;
 import me.hydos.lint.world.dimension.Dimensions;
 import net.minecraft.client.item.TooltipContext;
 import net.minecraft.entity.player.PlayerEntity;
@@ -97,7 +97,7 @@ public class WaypointTeleportItem extends Item {
         BlockPos pos = context.getBlockPos();
         World world = context.getWorld();
 
-        if (world.getBlockState(pos) == LintBlocks2.RETURN_HOME.getDefaultState().with(ReturnHomeBlock.ACTIVATED, true)) {
+        if (world.getBlockState(pos) == LintBlocks.RETURN_HOME.getDefaultState().with(ReturnHomeBlock.ACTIVATED, true)) {
             context.getStack().getOrCreateTag().putIntArray("waypoint", new int[]{pos.getX(), pos.getY() + 1, pos.getZ()});
             return ActionResult.SUCCESS;
         }
