+++ src/main/java/me/hydos/lint/item/group/ItemGroups.java
@@ -20,8 +20,8 @@
 package me.hydos.lint.item.group;
 
 import me.hydos.lint.Lint;
+import me.hydos.lint.block.LintBlocks;
 import me.hydos.lint.item.LintItems;
-import me.hydos.lint.refactord.block.LintBlocks2;
 import net.fabricmc.fabric.api.client.itemgroup.FabricItemGroupBuilder;
 import net.minecraft.item.ItemGroup;
 import net.minecraft.item.ItemStack;
@@ -30,12 +30,12 @@ public class ItemGroups {
 
     public static final ItemGroup BLOCKS = FabricItemGroupBuilder.create(
             Lint.id("lint_blocks"))
-            .icon(() -> new ItemStack(LintBlocks2.LIVELY_GRASS))
+            .icon(() -> new ItemStack(LintBlocks.LIVELY_GRASS))
             .build();
 
     public static final ItemGroup DECORATIONS = FabricItemGroupBuilder.create(
             Lint.id("lint_decorations"))
-            .icon(() -> new ItemStack(LintBlocks2.MYSTICAL_DAISY))
+            .icon(() -> new ItemStack(LintBlocks.MYSTICAL_DAISY))
             .build();
 
     public static final ItemGroup TOOLS = FabricItemGroupBuilder.create(
@@ -50,7 +50,7 @@ public class ItemGroups {
 
     public static final ItemGroup FOOD = FabricItemGroupBuilder.create(
             Lint.id("lint_food"))
-            .icon(() -> new ItemStack(LintBlocks2.COOKIE))
+            .icon(() -> new ItemStack(LintBlocks.COOKIE))
             .build();
 
     public static void register() {
