+++ src/main/java/me/hydos/lint/mixin/FeatureMixin.java
@@ -19,13 +19,14 @@
 
 package me.hydos.lint.mixin;
 
-import me.hydos.lint.refactord.block.DirtLikeBlock;
 import net.minecraft.block.Block;
 import net.minecraft.world.gen.feature.Feature;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+import me.hydos.lint.block.DirtLikeBlock;
 
 /**
  * @reason Make forest rocks generate.
