+++ src/main/java/me/hydos/lint/world/biome/Biomes.java
@@ -20,9 +20,9 @@
 package me.hydos.lint.world.biome;
 
 import me.hydos.lint.Lint;
+import me.hydos.lint.block.LintBlocks;
 import me.hydos.lint.entity.Birds;
 import me.hydos.lint.entity.Entities;
-import me.hydos.lint.refactord.block.LintBlocks2;
 import me.hydos.lint.sound.Sounds;
 import me.hydos.lint.world.biome.surface.DawnShardlandsEdgeSurfaceBuilder;
 import me.hydos.lint.world.biome.surface.DawnShardlandsSurfaceBuilder;
@@ -60,8 +60,8 @@ public class Biomes {
     /**
      * Biome Surface Builders
      */
-    public static final SurfaceBuilder<TernarySurfaceConfig> OCEAN_RAW_SB = new OceanSurfaceBuilder(LintBlocks2.MYSTICAL_SAND.getDefaultState(), LintBlocks2.WHITE_SAND.getDefaultState());
-    public static final SurfaceBuilder<TernarySurfaceConfig> CORRUPT_OCEAN_RAW_SB = new OceanSurfaceBuilder(LintBlocks2.CORRUPT_SAND.getDefaultState(), LintBlocks2.WHITE_SAND.getDefaultState());
+    public static final SurfaceBuilder<TernarySurfaceConfig> OCEAN_RAW_SB = new OceanSurfaceBuilder(LintBlocks.MYSTICAL_SAND.getDefaultState(), LintBlocks.WHITE_SAND.getDefaultState());
+    public static final SurfaceBuilder<TernarySurfaceConfig> CORRUPT_OCEAN_RAW_SB = new OceanSurfaceBuilder(LintBlocks.CORRUPT_SAND.getDefaultState(), LintBlocks.WHITE_SAND.getDefaultState());
     public static final SurfaceBuilder<TernarySurfaceConfig> DAWN_SHARDLANDS_RAW_SB = new DawnShardlandsSurfaceBuilder();
     public static final SurfaceBuilder<TernarySurfaceConfig> DAWN_SHARDLANDS_EDGE_RAW_SB = new DawnShardlandsEdgeSurfaceBuilder();
 
@@ -109,9 +109,9 @@ public class Biomes {
 
     public static final ConfiguredSurfaceBuilder<TernarySurfaceConfig> MF_SB = SurfaceBuilder.DEFAULT.withConfig(
             new TernarySurfaceConfig(
-                    LintBlocks2.LIVELY_GRASS.getDefaultState(),
-                    LintBlocks2.RICH_DIRT.getDefaultState(),
-                    LintBlocks2.RICH_DIRT.getDefaultState()));
+                    LintBlocks.LIVELY_GRASS.getDefaultState(),
+                    LintBlocks.RICH_DIRT.getDefaultState(),
+                    LintBlocks.RICH_DIRT.getDefaultState()));
 
     public static final Biome MYSTICAL_GROVE = new Biome.Builder()
             .precipitation(Biome.Precipitation.NONE)
@@ -193,9 +193,9 @@ public class Biomes {
             .build();
 
     private static final ConfiguredSurfaceBuilder<TernarySurfaceConfig> FF_SB = SurfaceBuilder.DEFAULT.withConfig(
-            new TernarySurfaceConfig(LintBlocks2.FROSTED_GRASS.getDefaultState(),
-                    LintBlocks2.RICH_DIRT.getDefaultState(),
-                    LintBlocks2.RICH_DIRT.getDefaultState()));
+            new TernarySurfaceConfig(LintBlocks.FROSTED_GRASS.getDefaultState(),
+                    LintBlocks.RICH_DIRT.getDefaultState(),
+                    LintBlocks.RICH_DIRT.getDefaultState()));
 
     public static final Biome ETHEREAL_WOODLAND = new Biome.Builder()
             .precipitation(Biome.Precipitation.NONE)
@@ -236,9 +236,9 @@ public class Biomes {
             .build();
 
     private static final ConfiguredSurfaceBuilder<TernarySurfaceConfig> CF_SB = SurfaceBuilder.DEFAULT.withConfig(
-            new TernarySurfaceConfig(LintBlocks2.CORRUPT_GRASS.getDefaultState(),
-                    LintBlocks2.RICH_DIRT.getDefaultState(),
-                    LintBlocks2.RICH_DIRT.getDefaultState()));
+            new TernarySurfaceConfig(LintBlocks.CORRUPT_GRASS.getDefaultState(),
+                    LintBlocks.RICH_DIRT.getDefaultState(),
+                    LintBlocks.RICH_DIRT.getDefaultState()));
     /**
      * Biomes
      */
@@ -280,9 +280,9 @@ public class Biomes {
 
     private static final ConfiguredSurfaceBuilder<TernarySurfaceConfig> OC_SB = OCEAN_RAW_SB.withConfig(
             new TernarySurfaceConfig(
-                    LintBlocks2.LIVELY_GRASS.getDefaultState(),
-                    LintBlocks2.RICH_DIRT.getDefaultState(),
-                    LintBlocks2.RICH_DIRT.getDefaultState()));
+                    LintBlocks.LIVELY_GRASS.getDefaultState(),
+                    LintBlocks.RICH_DIRT.getDefaultState(),
+                    LintBlocks.RICH_DIRT.getDefaultState()));
 
     public static final Biome OCEAN = new Biome.Builder()
             .precipitation(Biome.Precipitation.NONE)
@@ -321,9 +321,9 @@ public class Biomes {
             .build();
 
     private static final ConfiguredSurfaceBuilder<TernarySurfaceConfig> CB_SB = CORRUPT_OCEAN_RAW_SB.withConfig(
-            new TernarySurfaceConfig(LintBlocks2.CORRUPT_GRASS.getDefaultState(),
-                    LintBlocks2.RICH_DIRT.getDefaultState(),
-                    LintBlocks2.RICH_DIRT.getDefaultState()));
+            new TernarySurfaceConfig(LintBlocks.CORRUPT_GRASS.getDefaultState(),
+                    LintBlocks.RICH_DIRT.getDefaultState(),
+                    LintBlocks.RICH_DIRT.getDefaultState()));
 
     // TODO indigo stone layer in corrupt biomes
     public static final Biome CORRUPT_BEACH = new Biome.Builder()
@@ -358,9 +358,9 @@ public class Biomes {
             .build();
 
     private static final ConfiguredSurfaceBuilder<TernarySurfaceConfig> IN_SB = SurfaceBuilder.DEFAULT.withConfig(
-            new TernarySurfaceConfig(LintBlocks2.INDIGO_STONE.getDefaultState(),
-                    LintBlocks2.INDIGO_STONE.getDefaultState(),
-                    LintBlocks2.INDIGO_STONE.getDefaultState()));
+            new TernarySurfaceConfig(LintBlocks.INDIGO_STONE.getDefaultState(),
+                    LintBlocks.INDIGO_STONE.getDefaultState(),
+                    LintBlocks.INDIGO_STONE.getDefaultState()));
 
     public static final Biome INDIGO_RIDGES = new Biome.Builder()
             .precipitation(Biome.Precipitation.NONE)
@@ -394,9 +394,9 @@ public class Biomes {
             .build();
 
     private static final ConfiguredSurfaceBuilder<TernarySurfaceConfig> DS_SB = DAWN_SHARDLANDS_RAW_SB.withConfig(
-            new TernarySurfaceConfig(LintBlocks2.ASPHALT.getDefaultState(),
-                    LintBlocks2.ASPHALT.getDefaultState(),
-                    LintBlocks2.ASPHALT.getDefaultState()));
+            new TernarySurfaceConfig(LintBlocks.ASPHALT.getDefaultState(),
+                    LintBlocks.ASPHALT.getDefaultState(),
+                    LintBlocks.ASPHALT.getDefaultState()));
 
     public static final Biome DAWN_SHARDLANDS = new Biome.Builder()
             .precipitation(Biome.Precipitation.NONE)
@@ -423,9 +423,9 @@ public class Biomes {
             .build();
 
     private static final ConfiguredSurfaceBuilder<TernarySurfaceConfig> DSE_SB = DAWN_SHARDLANDS_EDGE_RAW_SB.withConfig(
-            new TernarySurfaceConfig(LintBlocks2.ASPHALT.getDefaultState(),
-                    LintBlocks2.ASPHALT.getDefaultState(),
-                    LintBlocks2.ASPHALT.getDefaultState()));
+            new TernarySurfaceConfig(LintBlocks.ASPHALT.getDefaultState(),
+                    LintBlocks.ASPHALT.getDefaultState(),
+                    LintBlocks.ASPHALT.getDefaultState()));
 
     public static final Biome DAWN_SHARDLANDS_EDGE = new Biome.Builder()
             .precipitation(Biome.Precipitation.NONE)
