+++ src/main/java/me/hydos/lint/world/biome/surface/DawnShardlandsEdgeSurfaceBuilder.java
@@ -19,7 +19,6 @@
 
 package me.hydos.lint.world.biome.surface;
 
-import me.hydos.lint.refactord.block.LintBlocks2;
 import net.minecraft.block.BlockState;
 import net.minecraft.world.biome.Biome;
 import net.minecraft.world.chunk.Chunk;
@@ -28,8 +27,10 @@ import net.minecraft.world.gen.surfacebuilder.TernarySurfaceConfig;
 
 import java.util.Random;
 
+import me.hydos.lint.block.LintBlocks;
+
 public class DawnShardlandsEdgeSurfaceBuilder extends SurfaceBuilder<TernarySurfaceConfig> {
-    private static final TernarySurfaceConfig ASH_CONFIG = new TernarySurfaceConfig(LintBlocks2.ASH.getDefaultState(), LintBlocks2.ASPHALT.getDefaultState(), LintBlocks2.ASH.getDefaultState());
+    private static final TernarySurfaceConfig ASH_CONFIG = new TernarySurfaceConfig(LintBlocks.ASH.getDefaultState(), LintBlocks.ASPHALT.getDefaultState(), LintBlocks.ASH.getDefaultState());
 
     public DawnShardlandsEdgeSurfaceBuilder() {
         super(TernarySurfaceConfig.CODEC);
