+++ src/main/java/me/hydos/lint/world/carver/LintCaveCarver.java
@@ -28,7 +28,7 @@ import org.apache.commons.lang3.mutable.MutableBoolean;
 import com.mojang.serialization.Codec;
 
 import me.hydos.lint.Lint;
-import me.hydos.lint.refactord.block.LintBlocks2;
+import me.hydos.lint.block.LintBlocks;
 import net.minecraft.block.BlockState;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.BlockPos.Mutable;
@@ -49,7 +49,7 @@ public class LintCaveCarver extends CaveCarver {
 
 	@Override
 	protected boolean canCarveBlock(BlockState state, BlockState stateAbove) {
-		return state.isOf(LintBlocks2.FUSED_STONE) || super.canCarveBlock(state, stateAbove);
+		return state.isOf(LintBlocks.FUSED_STONE) || super.canCarveBlock(state, stateAbove);
 	}
 
 	protected float getTunnelSystemWidth(Random random) {
@@ -68,7 +68,7 @@ public class LintCaveCarver extends CaveCarver {
 			pos1.set(x, y, z);
 			BlockState state = chunk.getBlockState(pos1);
 			BlockState upState = chunk.getBlockState(mutable2.set(pos1, Direction.UP));
-			if (state.isOf(LintBlocks2.CORRUPT_GRASS) || state.isOf(LintBlocks2.LIVELY_GRASS)) {
+			if (state.isOf(LintBlocks.CORRUPT_GRASS) || state.isOf(LintBlocks.LIVELY_GRASS)) {
 				grassCheckerThing.setTrue();
 			}
 
@@ -81,7 +81,7 @@ public class LintCaveCarver extends CaveCarver {
 					chunk.setBlockState(pos1, CAVE_AIR, false);
 					if (grassCheckerThing.isTrue()) {
 						pos3.set(pos1, Direction.DOWN);
-						if (chunk.getBlockState(pos3).isOf(LintBlocks2.RICH_DIRT)) {
+						if (chunk.getBlockState(pos3).isOf(LintBlocks.RICH_DIRT)) {
 							chunk.setBlockState(pos3, posToBiome.apply(pos1).getGenerationSettings().getSurfaceConfig().getTopMaterial(), false);
 						}
 					}
