+++ src/main/java/me/hydos/lint/world/feature/FadingAshFeature.java
@@ -21,7 +21,7 @@ package me.hydos.lint.world.feature;
 
 import java.util.Random;
 
-import me.hydos.lint.refactord.block.LintBlocks2;
+import me.hydos.lint.block.LintBlocks;
 import me.hydos.lint.world.gen.FraiyaTerrainGenerator;
 import net.minecraft.block.BlockState;
 import net.minecraft.util.math.BlockPos;
@@ -34,8 +34,8 @@ import net.minecraft.world.gen.feature.DefaultFeatureConfig;
 import net.minecraft.world.gen.feature.Feature;
 
 public class FadingAshFeature extends Feature<DefaultFeatureConfig> {
-	private static final BlockState ASH = LintBlocks2.ASH.getDefaultState();
-	private static final BlockState ASPHALT = LintBlocks2.ASPHALT.getDefaultState();
+	private static final BlockState ASH = LintBlocks.ASH.getDefaultState();
+	private static final BlockState ASPHALT = LintBlocks.ASPHALT.getDefaultState();
 
 	public FadingAshFeature() {
 		super(DefaultFeatureConfig.CODEC);
