+++ src/main/java/me/hydos/lint/world/feature/PortalFeature.java
@@ -21,8 +21,8 @@ package me.hydos.lint.world.feature;
 
 import java.util.Random;
 
-import me.hydos.lint.refactord.block.LintBlocks2;
-import me.hydos.lint.refactord.block.ReturnHomeBlock;
+import me.hydos.lint.block.LintBlocks;
+import me.hydos.lint.block.ReturnHomeBlock;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.PillarBlock;
 import net.minecraft.block.SlabBlock;
@@ -59,7 +59,7 @@ public class PortalFeature extends Feature<DefaultFeatureConfig> {
 		makePillar(world, pos.add(1, 0, 3));
 		makePillar(world, pos.add(3, 0, 3));
 		makeRoof(world, pos);
-		this.setBlockState(world, pos.add(2, 1, 2), LintBlocks2.RETURN_HOME.getDefaultState().with(ReturnHomeBlock.ACTIVATED, activated));
+		this.setBlockState(world, pos.add(2, 1, 2), LintBlocks.RETURN_HOME.getDefaultState().with(ReturnHomeBlock.ACTIVATED, activated));
 	}
 
 	private void makeBase(WorldAccess world, BlockPos pos) {
